OwnerIconを他のページでも扱えるようにファイルに抽出した
